deepfake-chain/backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
web3==6.11.3
python-dotenv==1.0.0
tensorflow
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.26.2
timm
torch
torchvision
scikit-learn==1.3.0
matplotlib==3.8.2